+++ src/FlysystemManager.php
@@ -14,6 +14,7 @@ namespace GrahamCampbell\Flysystem;
 use GrahamCampbell\Flysystem\Factories\FlysystemFactory;
 use GrahamCampbell\Manager\AbstractManager;
 use Illuminate\Contracts\Config\Repository;
+use InvalidArgumentException;
 
 /**
  * This is the flysystem manager class.
@@ -110,7 +111,7 @@ class FlysystemManager extends AbstractManager
         $connections = $this->config->get($this->getConfigName().'::connections');
 
         if (!is_array($config = array_get($connections, $name)) && !$config) {
-            throw new \InvalidArgumentException("Adapter [$name] not configured.");
+            throw new InvalidArgumentException("Adapter [$name] not configured.");
         }
 
         if (is_string($cache = array_get($config, 'cache'))) {
@@ -136,7 +137,7 @@ class FlysystemManager extends AbstractManager
         $cache = $this->config->get($this->getConfigName().'::cache');
 
         if (!is_array($config = array_get($cache, $name)) && !$config) {
-            throw new \InvalidArgumentException("Cache [$name] not configured.");
+            throw new InvalidArgumentException("Cache [$name] not configured.");
         }
 
         $config['name'] = $name;