+++ tests/Adapters/ConnectionFactoryTest.php
@@ -35,9 +35,11 @@ class ConnectionFactoryTest extends AbstractTestCase
     {
         return [
             ['awss3', 'AwsS3Connector'],
+            ['azure', 'AzureConnector'],
             ['copy', 'CopyConnector'],
             ['dropbox', 'DropboxConnector'],
             ['ftp', 'FtpConnector'],
+            ['gridfs', 'GridFSConnector'],
             ['local', 'LocalConnector'],
             ['null', 'NullConnector'],
             ['rackspace', 'RackspaceConnector'],